Add copyright headers, comment fixes, minor refactoring
